Cast stack sizes to size_t to fix errors on some compilers
The size of a stack trace is compiler defined, so depending on how the
compiler defines it, it might not be the same as the type of
m_max_trace. On my setup, g++ 14.2.1 x86_64-linux, m_stack_trace is a
long unsigned int (size_t) and stack.size() is a short unsigned int. std::min
requires that both arguments be the same type to properly instantiate
the template, so this produces a compiler error.

My fix is to cast the stack size to size_t. On my machine at least,
this is strictly bigger than it was, so this doesn't create any new
danger of overflow.