Fix opening blob cache folder on Linux
Looks like the previous code assumed that
std::filesystem::is_directory would return false for a path that
doesn't exist. On Linux, it appears to instead set the error code,
which causes JuniorsToolbox to print an error in the log instead of
just creating the directory. This fixes by checking whether the path
exists first, and creating if it doesn't, before checking if it's a
directory and giving an error if it isn't.