Don't fight the window manager
This code assumed that the user had exclusive control over the
size and position of all windows. While this is generally true of most
window managers, for users with a tiling window manager, this meant
that the code was constantly fighting the window manager for control,
causing the window to flash back and forth between multiple places.

The fix is to not assume that m_is_resized == (window->size !=
m_prev_size), and instead only try to set the size once when
m_is_resized is set to true, and the same for m_is_repositioned and positions.